+++ src/renderer/settings.js
@@ -1,16 +1,31 @@
 // @ts-nocheck
 const electron = window.nodeRequire("electron");
 const ipc = electron.ipcRenderer;
+var session;
 
 function initApp() {
   console.log("initApp");
   ipc.send("settings:loaded");
-  $("#tabs li").on("click", function() {
-    const tab = $(this).data("tab");
-    $("#tabs li").removeClass("is-active");
-    $(this).addClass("is-active");
-    $("#tab-content div.container").removeClass("is-active");
-    $("#tab-content div.container[data-content='" + tab + "']").addClass("is-active");
+  var editor = ace.edit("node-input-blacklist");
+  // editor.setTheme("ace/theme/tomorrow")
+  session = editor.getSession();
+  session.setMode('ace/mode/text')
+  session.on('change', function () {
+    $("#nodesexcludes").val(session.getValue());
+  });
+  $("#red-ui-tab-user-settings-nodered").click(function(){
+    $(this).addClass("active");
+    $("#user-settings-nodered").show();
+    $("#red-ui-tab-user-settings-other").removeClass("active");
+    $("#user-settings-other").hide();
+    return false;
+  });
+  $("#red-ui-tab-user-settings-other").click(function(){
+    $(this).addClass("active");
+    $("#user-settings-nodered").hide();
+    $("#red-ui-tab-user-settings-nodered").removeClass("active");
+    $("#user-settings-other").show();
+    return false;
   });
 }
 
@@ -55,9 +70,11 @@ ipc.on("settings:set", (event, settings) => {
   $("#allowprerelease").prop("checked", settings.allowPrerelease);
   $("#openlastfile").prop("checked", settings.openLastFile);
   $("#listenport").val(settings.listenPort);
+  session.setValue($("#nodesexcludes").val(),-1);
 })
 
 $("#button-submit").on("click", function(event) {
+  if (!$("#configForm").valid()) return true;
   const data = {
     userDir: $("#userdir").val(),
     credentialSecret: $("#credentialsecret").val(),
@@ -74,8 +91,27 @@ $("#button-submit").on("click", function(event) {
     listenPort: $("#listenport").val()
   }
   ipc.send("settings:update", data);
+  return false;
 });
 
 $("#button-cancel").on("click", function(event) {
   ipc.send("settings:cancel");
+  return false;
+});
+
+$("#configForm").validate({
+  rules: {
+    credentialsecret: {
+      required: true
+    },
+    listenport: {
+      range: [0, 65535]
+    },
+    httpnodeauthuser: {
+      maxlength: 64
+    },
+    httpnodeauthpass: {
+      maxlength: 64
+    }
+  }
 });